@@ -1,17 +1,16 @@
-﻿namespace Auditable.AspNetCore
+﻿namespace Auditable.AspNetCore;
+
+using Configuration;
+using Microsoft.Extensions.DependencyInjection;
+
+
+/// <summary>
+/// this is recommended way to register the ASPNET dependencies.
+/// </summary>
+public class AspNet : IExtension
 {
-    using Configuration;
-    using Microsoft.Extensions.DependencyInjection;
-
-
-    /// <summary>
-    /// this is recommended way to register the ASPNET dependencies.
-    /// </summary>
-    public class AspNet : IExtension
+    public void RegisterServices(IServiceCollection services)
     {
-        public void RegisterServices(IServiceCollection services)
-        {
-            services.AddAuditableForAspNetOnly();
-        }
+        services.AddAuditableForAspNetOnly();
     }
 }
