@@ -1,18 +1,17 @@
-﻿namespace Auditable.Collectors.Environment
-{
-    using System.Reflection;
-    using System.Threading.Tasks;
+﻿namespace Auditable.Collectors.Environment;
 
-    public class DefaultEnvironmentCollector : IEnvironmentCollector
+using System.Reflection;
+using System.Threading.Tasks;
+
+public class DefaultEnvironmentCollector : IEnvironmentCollector
+{
+    public Task<Environment> Extract()
     {
-        public Task<Environment> Extract()
+        var env = new Environment
         {
-            var env = new Environment
-            {
-                Application = Assembly.GetEntryAssembly().FullName,
-                Host = System.Environment.MachineName
-            };
-            return Task.FromResult(env);
-        }
+            Application = Assembly.GetEntryAssembly().FullName,
+            Host = System.Environment.MachineName
+        };
+        return Task.FromResult(env);
     }
 }
