@@ -1,8 +1,7 @@
-﻿namespace Auditable.Collectors.Environment
+﻿namespace Auditable.Collectors.Environment;
+
+public class Environment
 {
-    public class Environment
-    {
-        public string Host { get; set; }
-        public string Application { get; set; }
-    }
+    public string Host { get; set; }
+    public string Application { get; set; }
 }
