@@ -1,12 +1,11 @@
-﻿namespace Auditable.Collectors.Environment
-{
-    using System.Threading.Tasks;
+﻿namespace Auditable.Collectors.Environment;
 
-    /// <summary>
-    /// Grab information about the running environment
-    /// </summary>
-    public interface IEnvironmentCollector
-    {
-        Task<Environment> Extract();
-    }
+using System.Threading.Tasks;
+
+/// <summary>
+/// Grab information about the running environment
+/// </summary>
+public interface IEnvironmentCollector
+{
+    Task<Environment> Extract();
 }
