@@ -1,12 +1,11 @@
-﻿namespace Auditable.Collectors.Initiator
-{
-    using System.Threading.Tasks;
+﻿namespace Auditable.Collectors.Initiator;
 
-    /// <summary>
-    /// Grab information about the user who is acting the action
-    /// </summary>
-    public interface IInitiatorCollector
-    {
-        Task<Initiator> Extract();
-    }
+using System.Threading.Tasks;
+
+/// <summary>
+/// Grab information about the user who is acting the action
+/// </summary>
+public interface IInitiatorCollector
+{
+    Task<Initiator> Extract();
 }
