@@ -1,8 +1,7 @@
-﻿namespace Auditable.Collectors.Initiator
+﻿namespace Auditable.Collectors.Initiator;
+
+public class Initiator
 {
-    public class Initiator
-    {
-        public string Id { get; set; }
-        public string Name { get; set; }
-    }
+    public string Id { get; set; }
+    public string Name { get; set; }
 }
