@@ -1,12 +1,11 @@
-﻿namespace Auditable.Collectors.Request
-{
-    using System.Threading.Tasks;
+﻿namespace Auditable.Collectors.Request;
 
-    /// <summary>
-    /// grab information about the request (tracing)
-    /// </summary>
-    public interface IRequestContextCollector
-    {
-        Task<RequestContext> Extract();
-    }
+using System.Threading.Tasks;
+
+/// <summary>
+/// grab information about the request (tracing)
+/// </summary>
+public interface IRequestContextCollector
+{
+    Task<RequestContext> Extract();
 }
