@@ -1,9 +1,8 @@
-﻿namespace Auditable.Collectors.Request
+﻿namespace Auditable.Collectors.Request;
+
+public class RequestContext
 {
-    public class RequestContext
-    {
-        public string SpanId { get; set; }
-        public string TraceId { get; set; }
-        public string ParentId { get; set; }
-    }
+    public string SpanId { get; set; }
+    public string TraceId { get; set; }
+    public string ParentId { get; set; }
 }
