@@ -1,17 +1,16 @@
-﻿namespace Auditable.Configuration
+﻿namespace Auditable.Configuration;
+
+using Microsoft.Extensions.DependencyInjection;
+
+
+/// <summary>
+/// allows the extension to apply IoC setup/overrides
+/// </summary>
+public interface IExtension
 {
-    using Microsoft.Extensions.DependencyInjection;
-
-
     /// <summary>
-    /// allows the extension to apply IoC setup/overrides
+    /// internal to setup the dependencies and options
     /// </summary>
-    public interface IExtension
-    {
-        /// <summary>
-        /// internal to setup the dependencies and options
-        /// </summary>
-        /// <param name="services">the IoC setup</param>
-        void RegisterServices(IServiceCollection services);
-    }
+    /// <param name="services">the IoC setup</param>
+    void RegisterServices(IServiceCollection services);
 }
