@@ -1,9 +1,8 @@
-﻿namespace Auditable.Configuration
-{
-    using System;
+﻿namespace Auditable.Configuration;
 
-    public interface ISetupOptions<out T> where T : class, new()
-    {
-        void Setup(Action<T> options);
-    }
+using System;
+
+public interface ISetupOptions<out T> where T : class, new()
+{
+    void Setup(Action<T> options);
 }
