@@ -1,7 +1,6 @@
-﻿namespace Auditable.Configuration
-{
-    /// <summary>
-    /// setup a writer
-    /// </summary>
-    public interface IWriterProvider : IExtension { }
-}
+﻿namespace Auditable.Configuration;
+
+/// <summary>
+/// setup a writer
+/// </summary>
+public interface IWriterProvider : IExtension { }
