@@ -1,9 +1,8 @@
-﻿namespace Auditable.Delta
-{
-    using Newtonsoft.Json.Linq;
+﻿namespace Auditable.Delta;
 
-    public interface IDifferenceEngine
-    {
-        JToken Differences(string left, string right);
-    }
+using Newtonsoft.Json.Linq;
+
+public interface IDifferenceEngine
+{
+    JToken Differences(string left, string right);
 }
