@@ -1,7 +1,6 @@
-﻿namespace Auditable
+﻿namespace Auditable;
+
+public interface IAuditable
 {
-    public interface IAuditable
-    {
-        IAuditableContext CreateContext(string name, params object[] targets);
-    }
+    IAuditableContext CreateContext(string name, params object[] targets);
 }
