@@ -1,13 +1,12 @@
-﻿namespace Auditable
-{
-    using System;
-    using System.Threading.Tasks;
+﻿namespace Auditable;
 
-    public interface IAuditableContext : IDisposable, IAsyncDisposable
-    {
-        void WatchTargets(params object[] targets);
-        void Removed<T>(string id);
-        void Read<T>(string id);
-        Task WriteLog();
-    }
+using System;
+using System.Threading.Tasks;
+
+public interface IAuditableContext : IDisposable, IAsyncDisposable
+{
+    void WatchTargets(params object[] targets);
+    void Removed<T>(string id);
+    void Read<T>(string id);
+    Task WriteLog();
 }
