@@ -1,17 +1,16 @@
-﻿namespace Auditable.Infrastructure
+﻿namespace Auditable.Infrastructure;
+
+using Newtonsoft.Json;
+
+public class JsonSerializer
 {
-    using Newtonsoft.Json;
-
-    public class JsonSerializer
+    public string Serialize(object instance)
     {
-        public string Serialize(object instance)
-        {
-            return JsonConvert.SerializeObject(instance);
-        }
+        return JsonConvert.SerializeObject(instance);
+    }
 
-        public T Deserialize<T>(string payload)
-        {
-            return JsonConvert.DeserializeObject<T>(payload);
-        }
+    public T Deserialize<T>(string payload)
+    {
+        return JsonConvert.DeserializeObject<T>(payload);
     }
 }
