@@ -1,22 +1,21 @@
-﻿namespace Auditable.Infrastructure
+﻿namespace Auditable.Infrastructure;
+
+using System;
+
+public static class SystemDateTime
 {
-    using System;
+    private static Func<DateTime> _getDateTime = () => DateTime.UtcNow;
 
-    public static class SystemDateTime
+    public static DateTime UtcNow => _getDateTime();
+
+    public static void SetDateTime(Func<DateTime> getDateTime)
     {
-        private static Func<DateTime> _getDateTime = () => DateTime.UtcNow;
+        Code.Require(()=> getDateTime != null, nameof(getDateTime));
+        _getDateTime = getDateTime;
+    }
 
-        public static DateTime UtcNow => _getDateTime();
-
-        public static void SetDateTime(Func<DateTime> getDateTime)
-        {
-            Code.Require(()=> getDateTime != null, nameof(getDateTime));
-            _getDateTime = getDateTime;
-        }
-
-        public static void Reset()
-        {
-            _getDateTime = () => DateTime.UtcNow;
-        }
+    public static void Reset()
+    {
+        _getDateTime = () => DateTime.UtcNow;
     }
 }
