@@ -1,19 +1,18 @@
-﻿namespace Auditable.Parsing
+﻿namespace Auditable.Parsing;
+
+using Newtonsoft.Json;
+using Newtonsoft.Json.Converters;
+using Newtonsoft.Json.Linq;
+
+public class AuditableTarget
 {
-    using Newtonsoft.Json;
-    using Newtonsoft.Json.Converters;
-    using Newtonsoft.Json.Linq;
+    public string Type { get; set; }
+    public string Id { get; set; }
+    public JToken Delta { get; set; }
 
-    public class AuditableTarget
-    {
-        public string Type { get; set; }
-        public string Id { get; set; }
-        public JToken Delta { get; set; }
-
-        [JsonConverter(typeof(StringEnumConverter))] 
-        public ActionStyle Style { get; set; }
+    [JsonConverter(typeof(StringEnumConverter))] 
+    public ActionStyle Style { get; set; }
         
-        [JsonConverter(typeof(StringEnumConverter))]
-        public AuditType Audit { get; set; }
-    }
+    [JsonConverter(typeof(StringEnumConverter))]
+    public AuditType Audit { get; set; }
 }
