@@ -1,10 +1,9 @@
-﻿namespace Auditable.Parsing
-{
-    using System.Collections.Generic;
-    using System.Threading.Tasks;
+﻿namespace Auditable.Parsing;
 
-    public interface IParser
-    {
-        Task<string> Parse(string id, string actionName, IEnumerable<Target> targets);
-    }
+using System.Collections.Generic;
+using System.Threading.Tasks;
+
+public interface IParser
+{
+    Task<string> Parse(string id, string actionName, IEnumerable<Target> targets);
 }
