@@ -1,4 +1,3 @@
-﻿namespace Auditable.Writers.File
-{
-    public delegate string GetFileName(string auditId, string action);
-}
+﻿namespace Auditable.Writers.File;
+
+public delegate string GetFileName(string auditId, string action);
