@@ -1,9 +1,8 @@
-﻿namespace Auditable.Writers
-{
-    using System.Threading.Tasks;
+﻿namespace Auditable.Writers;
 
-    public interface IWriter
-    {
-        Task Write(string id, string action, string entry);
-    }
+using System.Threading.Tasks;
+
+public interface IWriter
+{
+    Task Write(string id, string action, string entry);
 }
